+++ server/src/main/java/org/realityforge/replicant/server/transport/ReplicantSessionManager.java
@@ -72,4 +72,9 @@ public interface ReplicantSessionManager
                           @Nullable String etag,
                           @Nonnull Collection<EntityMessage> messages,
                           @Nonnull ChangeSet changeSet );
+
+  /**
+   * Exposed so that bulk changes can reset Cache.
+   */
+  void deleteAllCacheEntries();
 }